+++ ltl2ba/buchi.c
@@ -113,7 +113,7 @@ int simplify_btrans() /* simplifies the transitions */
     getrusage(RUSAGE_SELF, &tr_fin);
     timeval_subtract (&t_diff, &tr_fin.ru_utime, &tr_debut.ru_utime);
     fprintf(tl_out, "\nSimplification of the Buchi automaton - transitions: %lli.%06ld",
-            (long long int)t_diff.tv_sec, t_diff.tv_usec);
+            (long long int)t_diff.tv_sec, (long int)t_diff.tv_usec);
     fprintf(tl_out, "\n%i transitions removed\n", changed);
 
   }
@@ -227,7 +227,7 @@ int simplify_bstates() /* eliminates redundant states */
     getrusage(RUSAGE_SELF, &tr_fin);
     timeval_subtract (&t_diff, &tr_fin.ru_utime, &tr_debut.ru_utime);
     fprintf(tl_out, "\nSimplification of the Buchi automaton - states: %lli.%06ld",
-            (long long int)t_diff.tv_sec, t_diff.tv_usec);
+            (long long int)t_diff.tv_sec, (long int)t_diff.tv_usec);
     fprintf(tl_out, "\n%i states removed\n", changed);
   }
 
@@ -672,7 +672,7 @@ void mk_buchi()
     getrusage(RUSAGE_SELF, &tr_fin);
     timeval_subtract (&t_diff, &tr_fin.ru_utime, &tr_debut.ru_utime);
     fprintf(tl_out, "\nBuilding the Buchi automaton : %lli.%06ld",
-            (long long int)t_diff.tv_sec, t_diff.tv_usec);
+            (long long int)t_diff.tv_sec, (long int)t_diff.tv_usec);
     fprintf(tl_out, "\n%i states, %i transitions\n", bstate_count, btrans_count);
   }
